src/reading_order/requirements.txt
networkx==3.0
numpy==1.24.2
Pillow==9.4.0
setuptools==65.6.3
tqdm==4.65.0
yaspin==2.3.0
